+++ R/RcppExports.R
@@ -9,10 +9,6 @@ quadProd <- function(a, b, kern, other_a, other_b, other_kern) {
     .Call(`_cort_quadProd`, a, b, kern, other_a, other_b, other_kern)
 }
 
-normMatrix <- function(ass, bs, kernels) {
-    .Call(`_cort_normMatrix`, ass, bs, kernels)
-}
-
 lossFunc <- function(bp, bin_repr, z) {
     .Call(`_cort_lossFunc`, bp, bin_repr, z)
 }